fix: Pass queryParams to the axios client (#12)
## Problem/motivation

I have some endpoints that depend on query params. When creating
clients, it looks like these params are not passed on to the axios
client and thus are never reaching the server.

## Solution

Pass queryParams on as params, as documented by one of the comments:

```
    // Create the actual URL by substituting params (if any) into the path pattern.
    // NB: what axios calls `params` are really queryparams, so different from our `params`,
```

Note: the tests weren't passing for me, but `client-server.test.ts`
looks like the right place to add a test case for this. Maybe HEAD is
broken or it's just my envo?

A test cases has been added, the whole suite should pass after #13.

## Follow-ups

The cast required in the route handler (`req.query.alsoAdd as string`)
does indicate there may be some type definition stuff that could be
added to the query param system?

---------

Co-authored-by: Sam Becker <820219+Sam152@users.noreply.github.com>
Co-authored-by: Anthony Manning-Franklin <anthony@workitspaces.com.au>